Refactored Interpreter
