+++ src/RamSynthesiser.h
@@ -16,23 +16,42 @@
 
 #pragma once
 
-#include "RamExecutor.h"
+#include "RamData.h"
+#include "RamProgram.h"
+#include "RamRelation.h"
+#include "SymbolTable.h"
 
+#include "RamStatement.h"
+
+#include <ostream>
+#include <vector>
 #include <string>
 
 namespace souffle {
 
 /**
- * A RAM executor based on the creation and compilation of an executable conducting
- * the actual computation.
+ * A RAM synthesiser: synthesises a C++ program from a RAM program.
  */
-class RamSynthesiser : public RamExecutor {
+class RamSynthesiser {
 private:
+    /** An optional stream to print logging information to an output stream */
+    std::ostream* report;
+
+    /** compile command */
     std::string compileCmd;
 
 public:
+
+    /**
+     * Updates logging stream
+     */
+    void setReportTarget(std::ostream& report) {
+        this->report = &report;
+    }
+
+public:
     /** A simple constructor */
-    RamSynthesiser(const std::string& compileCmd) : compileCmd(compileCmd) {}
+    RamSynthesiser(const std::string& compileCmd) : report(nullptr), compileCmd(compileCmd) {}
 
     /**
      * Generates the code for the given ram statement.The target file
@@ -60,20 +79,6 @@ public:
     std::string executeBinary(const SymbolTable& symTable, const RamProgram& prog,
             const std::string& filename = "", const int index = -1) const;
 
-    /**
-     * The actual implementation of this executor encoding the given
-     * program into a source file, compiling and executing it.
-     */
-    void applyOn(const RamProgram& prog, RamEnvironment& env, RamData* data) const override;
-
-    /**
-     * Execute sub-routines
-     */
-    virtual void executeSubroutine(RamEnvironment& env, const RamStatement& stmt,
-            const std::vector<RamDomain>& arguments, std::vector<RamDomain>& returnValues,
-            std::vector<bool>& returnErrors) const override {
-        // nothing to do here
-    }
 };
 
 }  // end of namespace souffle
